ATR-762: moved PX1028 unit tests to DAC declaration tests
